final-pjt-back/requirements.txt
asgiref==3.3.4
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.2
django-allauth==0.50.0
django-cors-headers==3.13.0
django-seed==0.2.2
djangorestframework==3.12.4
Faker==8.1.1
idna==3.4
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.12
